==> 1_APA_Project_clean.ipynb <==
# **APA Project**

**Read in data**

import pandas as pd
data = pd.read_csv('players_22.csv')
data.head(3)

data.shape

# **Data Cleaning**

#Drop Goalkeepers
data.drop(data[data['player_positions'] == 'GK'].index, inplace=True)

#Remove columns with external links - no use in machine learning
data = data[data.columns.drop(list(data.filter(regex='url')))]

#Drop columns where target variable is not stored
data = data.dropna(subset=['value_eur'])

#Drop columns not needed: jerseys, names, goalkeeping statistics, club/nationality stats, id, real face in game
data = data.drop(['long_name', 'dob', 'nation_jersey_number', 'short_name', 'club_jersey_number', 'sofifa_id', 'real_face', 'gk',
 'goalkeeping_diving','goalkeeping_handling','goalkeeping_kicking','goalkeeping_positioning','goalkeeping_reflexes',
 'goalkeeping_speed', 'nationality_id', 'player_tags', 'player_traits', 'club_name', 'club_loaned_from', 'club_joined',
  'league_name', 'player_positions', 'nationality_name', 'body_type'  ], axis=1)
data.head()

#Identify positional columns
data.columns[50:100]

#Drop positional columns
data = data.loc[:, :'defending_sliding_tackle']

# View unique dtypes of all variables
unique_dtypes = data.dtypes.unique()
unique_dtypes.tolist()

#Identify non-numeric cols for encoding
sel_cols = list(data.select_dtypes(include='O'))
print(sel_cols)

**Missing Data**

data.shape

data.isna().sum()

data.isnull().sum().sum()

#National team information: As data is missing for 16434/17041(96% of values) decision is made to drop this information
data.drop('nation_team_id', inplace=True, axis=1)

data.drop('nation_position', inplace=True, axis=1)

data.isna().sum()

**Encoding Categorical Variables**

Encoding Preferred Foot

#Preferred Foot
data['preferred_foot'].unique()

from sklearn.preprocessing import LabelEncoder
le = LabelEncoder()
data['preferred_foot'] = le.fit_transform(data['preferred_foot'])

#Preferred Foot
data['preferred_foot'].unique()

Encoding Position

#Position Definitions
att = ['RW', 'RS', 'ST' , 'LW',  'CF', 'LS',  'LF',  'RF']
mid = ['RCM',  'CDM',  'RDM', 'LCM',  'LDM', 'CM','RM', 'CAM', 'LM',  'LAM', 'RAM']
defe = [ 'LCB', 'LWB', 'LB', 'RCB', 'RB', 'CB', 'RWB']

data['club_position'].unique()

#Put Positions into 3 categories - Club Position col 9
new_positions = []
index = 0
for row in data.iterrows():
  x = (data.iloc[index, 9])
  if x == 'RW' or x == 'RS' or x == 'ST' or x == 'LW' or  x == 'CF' or x == 'LS' or  x == 'LF' or  x == 'RF':
      new_positions.append('ATT')
  elif x == 'RCM' or  x == 'CDM' or  x == 'RDM' or x == 'LCM' or  x == 'LDM' or x == 'CM' or x == 'RM' or x == 'CAM' or x == 'LM' or  x == 'LAM' or x == 'RAM':
      new_positions.append('MID')
  elif x == 'LCB' or x == 'LWB' or x == 'LB' or x == 'RCB' or x == 'RB' or x == 'CB' or x == 'RWB':
      new_positions.append('DEF')
  else:
      new_positions.append('other')
  index = index + 1
new_club_positions = new_positions

data['club_position'] = new_club_positions

from sklearn.preprocessing import LabelEncoder
le = LabelEncoder()
data['club_position'] = le.fit_transform(data['club_position'])#
data['club_position'].unique()

#Identify non-numeric cols for encoding
sel_cols = list(data.select_dtypes(include='O'))
print(sel_cols)

Encoding Work Rate

data['work_rate'].unique()

#Split into two columsn
data[['att_wr', 'def_wr']] = data['work_rate'].str.split('/', expand=True)
data['att_wr']

data['def_wr']

data['att_wr'].unique()

data['def_wr'].unique()

from sklearn.preprocessing import LabelEncoder
le = LabelEncoder()
data['att_wr'] = le.fit_transform(data['att_wr'])

from sklearn.preprocessing import LabelEncoder
le = LabelEncoder()
data['def_wr'] = le.fit_transform(data['def_wr'])

#Drop non-numeric col
data.drop('work_rate', inplace=True, axis=1)

#Identify non-numeric cols for encoding
sel_cols = list(data.select_dtypes(include='O'))
print(sel_cols)

# View unique dtypes of all variables
# Ensure all numeric
unique_dtypes = data.dtypes.unique()
unique_dtypes.tolist()

data.head()

data.columns[data.isna().any()].tolist()

Drop all rows with target variable as missing.

data.dropna(subset=['release_clause_eur'],inplace=True)

print(data.isnull().values.any())

data.columns[data.isna().any()].tolist()

data.isnull().sum().sum()

data

from google.colab import files
data.to_csv('final_clean_data.csv', index=None)
files.download("final_clean_data.csv")
