==> 3_APA_Shuffle_Split_data.ipynb <==
Dataset shuffling

#Read in previously cleaned and scaled data
import pandas as pd
data = pd.read_csv('final_scaled_data.csv', index_col=False)
data

data.drop(data.tail(1).index,inplace=True) 

print(data.isnull().values.any())

Train Test Split

from sklearn.model_selection import train_test_split

X=data.drop("value_eur", axis=1)
Y=data["value_eur"]
X

Y

#X_VAL = x_validation, same as x_test
X_train, X_val, y_train, y_val = train_test_split(X,
                                                  Y,
                                                  test_size=0.10,
                                                  shuffle=True,
                                                  random_state=324
                                                 )

X_train.shape

y_train.shape

X_val.shape

y_val.shape

X_train.head(3)

X_val.head(3)

y_train.head(3)

y_val.head(3)

print(X_train.isnull().values.any())

print(len(X_train))

print(len(X_val))

print(len(y_train))

print(len(y_val))

print(X_val.isnull().values.any())

print(y_train.isnull().values.any())

print(y_val.isnull().values.any())

Download files

from google.colab import files
X_train.to_csv('final_APA_X_train.csv', index=None)
X_val.to_csv('final_APA_X_val.csv', index=None)
y_train.to_csv('final_APA_y_train.csv', index=None)
y_val.to_csv('final_APA_y_val.csv', index=None)
files.download("final_APA_X_train.csv")
files.download("final_APA_X_val.csv")
files.download("final_APA_y_train.csv")
files.download("final_APA_y_val.csv")

files.download("final_APA_X_train.csv")
