==> 5_APA_Models_on_best_5_vars.ipynb <==
import pandas as pd
x_train = pd.read_csv('final_APA_X_train.csv')
x_test = pd.read_csv('final_APA_X_val.csv')
y_train = pd.read_csv('final_APA_y_train.csv')
y_test = pd.read_csv('final_APA_y_val.csv')

from pandas.core.window.rolling import numba_notes
#Regression Models
from sklearn.ensemble import GradientBoostingRegressor as XGBRegressor
#Linear Regression ( https://www.w3schools.com/python/python_ml_multiple_regression.asp )
from sklearn.linear_model import LinearRegression
lm = LinearRegression()
#Ridge Regression ( https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Ridge.html )
from sklearn import linear_model
ridge = linear_model.Ridge()
#Lasso Regression ( https://towardsdatascience.com/hyperparameter-tuning-in-lasso-and-ridge-regressions-70a4b158ae6d )
from sklearn.linear_model import Lasso, LassoCV, LinearRegression
lasso = linear_model.Lasso()
#KNeighborsRegressor
from sklearn.neighbors import KNeighborsRegressor
neigh = KNeighborsRegressor()
#Random Forrest ( https://towardsdatascience.com/random-forest-in-python-24d0893d51c0 )
from sklearn.ensemble import RandomForestRegressor
rf = RandomForestRegressor()
#Decision Trees ( https://medium.com/pursuitnotes/decision-tree-regression-in-6-steps-with-python-1a1c5aa2ee16 )
from sklearn.tree import DecisionTreeRegressor
dt = DecisionTreeRegressor()
#SVM - SVR
from sklearn import svm
SVR_m = svm.SVR()
#Gradient Boosting Regressor ( https://scikit-learn.org/stable/modules/generated/sklearn.ensemble.GradientBoostingRegressor.html#sklearn.ensemble.GradientBoostingRegressor )
from sklearn.ensemble import GradientBoostingRegressor
gbr = GradientBoostingRegressor()
# create an xgboost regression model: higher performance than gradient boosting
import xgboost as xg
xgb = xg.XGBRegressor()

#Best models
from sklearn.svm import SVR

#Linear regression - no tuning
model_1 = LinearRegression()
#Ridge Regression
model_2 = linear_model.Ridge(alpha=0.0)
#Lasso Regression
model_3 = linear_model.Lasso(alpha=0.0)
#K Nearest Neighbours Regression
model_4 = KNeighborsRegressor(n_neighbors = 10, weights= 'distance')
#Random Forrest Regressor
model_5 = RandomForestRegressor(max_depth=5, n_estimators=550)
#DT Regressors
model_6 = DecisionTreeRegressor(max_depth=3, max_features='auto', min_weight_fraction_leaf=0.2)
#SVM Regressor
model_7 = SVR(C=10, gamma=0.001)
#Gradient Boosting Regressor 
model_8 = GradientBoostingRegressor(learning_rate=0.02, n_estimators=80, random_state=0)
#XGBoost
model_9 = xg.XGBRegressor(n_estimators=60, max_depth = 3, learning_rate = 0.2)

models = [model_1, model_2, model_3, model_4, model_5, model_6, model_7, model_8, model_9]

for i in models:
  print(i)

selected_features = ['overall', 'wage_eur', 'release_clause_eur', 'attacking_short_passing', 'international_reputation']

from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
from math import sqrt

x_train

x_test

y_train

y_test

x_train.columns

x_train_best_features = x_train[['overall', 'wage_eur', 'release_clause_eur', 'attacking_short_passing', 'international_reputation']].copy()
x_test_best_features = x_test[['overall', 'wage_eur', 'release_clause_eur', 'attacking_short_passing', 'international_reputation']].copy()

x_train_best_features

x_test_best_features

x_train_best_features.isna().sum()

x_test_best_features.isna().sum()

#Run each model 
model_eval = []
for model in models:
  model.fit(x_train_best_features, y_train)
  predictions = model.predict(x_test_best_features)
  type(predictions)
  #FIX CALCS
  # Calculate the absolute errors
  errors = abs(predictions - y_test.values)
  # Calculate model evaluation metrics
  mae = mean_absolute_error(y_test, predictions)
  mse = mean_squared_error(y_test, predictions)
  rmse = sqrt(mse)
  r2 = r2_score(y_test, predictions)
  print(f'MODEL: {model}')
  print('Mean Absolute Error: ', mae, ' degrees.')
  print('Mean Squared Error:', mse, ' degrees')
  print('RMSE:', rmse, ' degrees')
  print('R2:', r2)
  model_eval.append({'Model': model, 'Mean Absolute Error': mae, 'Mean Squared Error': mse, 'RMSE': rmse,'R-squared': r2})

df = pd.DataFrame(model_eval)

df

from google.colab import files
df.to_csv('top_5_vars_results.csv', index=None)
files.download("top_5_vars_results.csv")
